DATACMNS-489 - Support for Future<T> as return value wrapper for repository methods.
Added a converter to process a NullableWrapper into a Future by producing an AsyncResult. This is to support repository methods annotated with @Async and returning a Future<T>. With Spring's asynchronous method invocation support activated this will cause the repository method being executed asynchronously.

During the implementation we stumbled above SPR-11715 which seems to prevent the repository methods being considered eligible for async proxying as AbstractAdvisingBeanPostProcessor inspects the proxy target class only. I've changed the factory setup to create opaque proxies for the time being to get the test for this special feature running. However, this breaks other tests and has to be removed before merging the feature back into master.

Related tickets: DATACMNS-483, SPR-11715.